Fix linting

+++ fmm-llvm/src/expression.rs
@@ -382,7 +382,7 @@ fn compile_size_of<'c>(context: &Context<'c>, size_of: &SizeOf) -> inkwell::valu
         context,
         context
             .target_data()
-            .get_store_size(&type_::compile(context, size_of.type_())) as u64,
+            .get_store_size(&type_::compile(context, size_of.type_())),
     )
 }
 
@@ -424,7 +424,7 @@ fn compile_primitive<'c>(
         Primitive::Float64(number) => context
             .inkwell()
             .f64_type()
-            .const_float(number as f64)
+            .const_float(number)
             .into(),
         Primitive::Integer8(number) => context
             .inkwell()
